--- app/src/main/assets/models/etanoteste.mtl
@@ -1,2 +1,0 @@
-# Blender 4.3.2 MTL File: 'etanoteste.blend'
-# www.blender.org
