--- app/src/main/assets/models/etanoteste.obj
@@ -1,4 +1,0 @@
-# Blender 4.3.2
-# www.blender.org
-mtllib etanoteste.mtl
-o oi
