--- app/src/main/java/com/google/ar/core/examples/java/principal/rendering/CompostoService.java
@@ -1,14 +1,0 @@
-package com.google.ar.core.examples.java.principal.rendering;
-
-import retrofit2.Call;
-import retrofit2.http.GET;
-import retrofit2.http.Path;
-
-public interface CompostoService {
-
-    @GET("compostos/formato/{formato}")
-    Call<Composto> getCompostoByFormato(@Path("formato") String formato);
-
-    @GET("compostos/nomenclatura/{nomenclatura}")
-    Call<Composto> getCompostoByNomenclatura(@Path("nomenclatura") String nomenclatura);
-}
