--- app/src/main/java/com/google/ar/core/examples/java/principal/rendering/validation/OnConnectionCheckListener.java
@@ -1,5 +1,0 @@
-package com.google.ar.core.examples.java.principal.rendering.validation;
-
-public interface OnConnectionCheckListener {
-    void onConnectionChecked(boolean isConnected);
-}
