--- app/src/main/java/com/google/ar/core/examples/java/principal/rendering/validation/OnDatabaseResultListener.java
@@ -1,6 +1,0 @@
-package com.google.ar.core.examples.java.principal.rendering.validation;
-
-public interface OnDatabaseResultListener {
-    void onSuccess(String object3DPath, String config3DPath, String texturaPath);
-    void onError(Exception e);
-}
