+++ app/src/main/java/config/principal/rendering/CompostoService.java
@@ -1,0 +1,14 @@
+package config.principal.rendering;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+import retrofit2.http.Path;
+
+public interface CompostoService {
+
+    @GET("compostos/formato/{formato}")
+    Call<Composto> getCompostoByFormato(@Path("formato") String formato);
+
+    @GET("compostos/nomenclatura/{nomenclatura}")
+    Call<Composto> getCompostoByNomenclatura(@Path("nomenclatura") String nomenclatura);
+}
