+++ app/src/main/java/config/principal/rendering/listeners/OnConnectionCheckListener.java
@@ -1,0 +1,5 @@
+package config.principal.rendering.listeners;
+
+public interface OnConnectionCheckListener {
+    void onConnectionChecked(boolean isConnected);
+}
