+++ app/src/main/java/config/principal/rendering/listeners/OnDatabaseResultListener.java
@@ -1,0 +1,6 @@
+package config.principal.rendering.listeners;
+
+public interface OnDatabaseResultListener {
+    void onSuccess(String object3DPath, String config3DPath, String texturaPath);
+    void onError(Exception e);
+}
