+++ app/src/main/java/config/principal/rendering/listeners/TextRecognitionListener.java
@@ -1,0 +1,7 @@
+package config.principal.rendering.listeners;
+
+public interface TextRecognitionListener {
+    void onObjectFound(String object3DPath, String config3DPath, String texturePath);
+    void onTextRecognitionFailed(Exception e);
+
+}
