+++ app/src/main/res/layout/activity_main.xml
@@ -21,7 +21,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.google.ar.core.examples.java.principal.QuimicAR">
+    tools:context="config.principal.QuimicAR">
 
   <android.opengl.GLSurfaceView
       android:id="@+id/surfaceview"
